requirements.txt
python-dotenv
pyTelegramBotAPI
qbittorrent-api
